+++ diario/Cursos/manifest.mf
