+++ diario/Cursos/nbproject/genfiles.properties
