+++ diario/Cursos/nbproject/project.properties
@@ -3,9 +3,8 @@ annotation.processing.enabled.in.editor=false
 annotation.processing.processors.list=
 annotation.processing.run.all.processors=true
 annotation.processing.source.output=${build.generated.sources.dir}/ap-source-output
-application.title=app
-application.vendor=ymegam
-auxiliary.org-netbeans-spi-editor-hints-projects.perProjectHintSettingsFile=nbproject/cfg_hints.xml
+application.title=TrabalhoFinalDesktop
+application.vendor=samue
 build.classes.dir=${build.dir}/classes
 build.classes.excludes=**/*.java,**/*.form
 # This directory is removed when the project is cleaned:
@@ -30,18 +29,18 @@ debug.test.modulepath=\
     ${run.test.modulepath}
 # This directory is removed when the project is cleaned:
 dist.dir=dist
-dist.jar=${dist.dir}/app.jar
+dist.jar=${dist.dir}/TrabalhoFinalDesktop.jar
 dist.javadoc.dir=${dist.dir}/javadoc
 endorsed.classpath=
 excludes=
-file.reference.java.app-lib=lib
+file.reference.mysql-connector-java-8.0.18.jar=C:\\Users\\Aluno\\Downloads\\mysql-connector-java-8.0.18\\mysql-connector-java-8.0.18.jar
 includes=**
 # Non-JavaFX jar file creation is deactivated in JavaFX 2.0+ projects
 jar.archive.disabled=true
 jar.compress=false
 javac.classpath=\
-    ${file.reference.java.app-lib}:\
-    ${javafx.classpath.extension}
+    ${javafx.classpath.extension}:\
+    ${file.reference.mysql-connector-java-8.0.18.jar}
 # Space-separated list of extra javac options
 javac.compilerargs=
 javac.deprecation=false
@@ -77,25 +76,20 @@ javafx.classpath.extension=\
     ${platforms.JDK_8.home}/jre/lib/javaws.jar:\
     ${platforms.JDK_8.home}/jre/lib/deploy.jar:\
     ${platforms.JDK_8.home}/jre/lib/plugin.jar
-javafx.deploy.adddesktopshortcut=false
-javafx.deploy.addstartmenushortcut=false
 javafx.deploy.allowoffline=true
 # If true, application update mode is set to 'background', if false, update mode is set to 'eager'
 javafx.deploy.backgroundupdate=false
-javafx.deploy.disable.proxy=false
 javafx.deploy.embedJNLP=true
 javafx.deploy.includeDT=true
-javafx.deploy.installpermanently=false
-javafx.deploy.permissionselevated=false
-# Set true to prevent creation of temporary copy of deployment artifacts before each run (disables concurrent runs)
+# Definir como verdadeiro impede a cria\u00e7\u00e3o de c\u00f3pia tempor\u00e1ria de artefatos de implanta\u00e7\u00e3o antes de cada execu\u00e7\u00e3o (desativa as execu\u00e7\u00f5es atuais)
 javafx.disable.concurrent.runs=false
-# Set true to enable multiple concurrent runs of the same WebStart or Run-in-Browser project
+# Definir como verdadeiro ativa v\u00e1rias execu\u00e7\u00f5es simult\u00e2neas do mesmo WebStart ou projeto Run-in-Browser
 javafx.enable.concurrent.external.runs=false
 # This is a JavaFX project
 javafx.enabled=true
 javafx.fallback.class=com.javafx.main.NoJavaFXFallback
 # Main class for JavaFX
-javafx.main.class=app.MainApp
+javafx.main.class=Principal.Main
 javafx.preloader.class=
 # This project does not use Preloader
 javafx.preloader.enabled=false
@@ -107,26 +101,21 @@ javafx.preloader.type=none
 javafx.rebase.libs=false
 javafx.run.height=600
 javafx.run.width=800
-javafx.signing.blob=false
-javafx.signing.enabled=false
-javafx.signing.type=notsigned
 jlink.launcher=false
-jlink.launcher.name=app
+jlink.launcher.name=TrabalhoFinalDesktop
 # Pre-JavaFX 2.0 WebStart is deactivated in JavaFX 2.0+ projects
 jnlp.enabled=false
 # Main class for Java launcher
 main.class=com.javafx.main.Main
-# For improved security specify narrower Codebase manifest attribute to prevent RIAs from being repurposed
+# Para uma melhor seguran\u00e7a, especifique um atributo de manifesto da Base de c\u00f3digo mais restrito para impedir que RIAs sejam adaptados
 manifest.custom.codebase=*
-# Specify Permissions manifest attribute to override default (choices: sandbox, all-permissions)
+# Especificar o atributo de manifesto de Permiss\u00f5es para substituir o default (op\u00e7\u00f5es: sandbox, all-permissions)
 manifest.custom.permissions=
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
-native.bundling.enabled=false
 platform.active=JDK_8
 run.classpath=\
-    ${dist.jar}:\
     ${javac.classpath}:\
     ${build.classes.dir}
 run.modulepath=\
@@ -137,6 +126,5 @@ run.test.classpath=\
 run.test.modulepath=\
     ${javac.test.modulepath}
 source.encoding=UTF-8
-src.java.dir=src/java
-src.resources.dir=src/resources
+src.dir=src
 test.src.dir=test
