+++ diario/Cursos/src/Principal/DbConnector.java
@@ -15,7 +15,7 @@ import java.sql.SQLException;
  */
 public class DbConnector {
     public static Connection getConnection() throws SQLException{
-        Connection con = DriverManager.getConnection("jdbc:mysql://localhost:/diario?useTimezone=true&serverTimezone=UTC","root","123456");
+        Connection con = DriverManager.getConnection("jdbc:mysql://localhost:/diario?useTimezone=true&serverTimezone=UTC","root","");
         
         return con;
     }
