+++ diario/Cursos/src/Principal/style.css
@@ -1,4 +1,4 @@
-
+@import url('https://fonts.googleapis.com/css?family=Montserrat&display=swap');
 
 table-view, .tree-table-view {
     /* Constants used throughout the tableview. */
@@ -61,4 +61,11 @@ table-view, .tree-table-view {
     -fx-background-color: #2C46A5;
     -fx-text-fill: white;
 
+}
+.coluna{
+    -fx-font-family: 'Montserrat',sans-serif;
+}
+
+.fonte{
+    -fx-font-family: 'Montserrat',sans-serif;
 }
